agents/requirements.txt
aenum==3.1.15
aioboto3==14.3.0
aiobotocore==2.22.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-retry==2.9.1
aioice==0.10.1
aioitertools==0.12.0
aioredis==2.0.1
aiortc==1.11.0
aiosignal==1.3.2
amazon-transcribe==0.6.4
annotated-types==0.7.0
annoy==1.17.3
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
av==14.2.0
awscrt==0.26.1
boto3==1.37.3
botocore==1.37.3
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
cryptography==44.0.2
dataclasses-json==0.6.7
deepgram-sdk==3.10.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
eval_type_backport==0.2.2
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.3.0
google-api-core==2.24.2
google-auth==2.40.2
google-cloud-speech==2.32.0
google-cloud-texttospeech==2.27.0
google-crc32c==1.7.1
google-genai==1.17.0
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
humanfriendly==10.0
idna==3.10
ifaddr==0.2.0
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
livekit==1.0.8
livekit-agents==1.0.22
livekit-api==1.0.2
livekit-plugins-assemblyai==1.0.11
livekit-plugins-aws==1.0.22
livekit-plugins-cartesia==1.0.11
livekit-plugins-deepgram==1.0.11
livekit-plugins-elevenlabs==1.0.22
livekit-plugins-google==1.0.22
livekit-plugins-neuphonic==1.0.22
livekit-plugins-noise-cancellation==0.2.1
livekit-plugins-openai==1.0.11
livekit-plugins-rag==1.0.13
livekit-plugins-silero==1.0.11
livekit-plugins-turn-detector==1.0.11
livekit-protocol==1.0.2
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.2.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==2.2.4
onnxruntime==1.21.0
openai==1.69.0
packaging==24.2
pillow==11.1.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pyee==13.0.0
PyJWT==2.10.1
pylibsrtp==0.12.0
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==6.1.0
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
s3transfer==0.11.3
safetensors==0.5.3
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
starlette==0.46.1
sympy==1.13.3
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.50.3
twilio==9.5.2
types-protobuf==4.25.0.20240417
typing==3.7.4.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.1
watchfiles==1.0.5
websockets==14.2
wrapt==1.17.2
yarl==1.18.3
